Add edit quiz capability, add list of user quizzes.

+++ src/main/java/com/gusev/spring5quizapp/repository/QuizRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface QuizRepository extends CrudRepository<Quiz, Integer> {
+public interface QuizRepository extends CrudRepository<Quiz, Long> {
 
     List<Quiz> findByTag(String tag);
